break internal package dependency cycle
