+++ packages/shared/package.json
@@ -142,7 +142,6 @@
   "devDependencies": {
     "@swc/jest": "^0.2.24",
     "@tamanu/build-tooling": "*",
-    "@tamanu/fake-data": "*",
     "chai": "^4.1.2",
     "concurrently": "^9.2.1",
     "jest": "^29.3.1",
