+++ packages/shared/src/test-helpers/index.js
@@ -1,6 +1,4 @@
 export * from './errors';
-export * from './factory';
 export * from './permissions';
 export * from './rawFiles';
-export * from './roles';
 export * from './sync';
